backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
email-validator
python-multipart
passlib[bcrypt]
bcrypt<4.0.0
python-jose
sqlalchemy
alembic
psycopg2-binary
python-dotenv
pytest
httpx
google-auth
